Refactor proxy logic out of Resource subclasses
The whole proxy idea is still kind of convoluted. Maybe one or two more passes
at it might help. The alternative is to get rid of proxies altogether and pass
the client into each Resource instance, and have the instance handle its own
object cache... or is there something else that's better?