Fix Participants instance resource creation
I realized we already had a test for muting participants. However, the mock
JSON endpoint was incorrect - it assumed Participants have a 'sid', while they
actually have a call_sid that you are supposed to use for querying.

This changes the mock JSON to do the right thing, and also overrides
getObjectFromJson to take a parameter in case the id parameter is not named
'sid' but instead named something silly like 'call_sid'. Also deletes the
duplicate test and fixes the old, broken test.