Don't update the URI when updating a resource
An InstanceResource would overwrite its URI if properties were retrieved or
updated on the object. However the URI returned from the API already has
'.json' appended to it, so when the library tries to also append .json, the
extension has been appended twice.

This is a similar bug to https://github.com/twilio/twilio-php/pull/71.
I have checked to try and make sure it cannot appear under any other
circumstances in the library.