Add env var to write raw HTTP traffic to stderr
It can be difficult to see the raw traffic going over the wire, but that's what
you need to write tests that integrate with the twilio-php client - the kind
you use for all of your Holodeck tests!

This isn't *exactly* what goes over the wire, but it's close enough for the
purposes of testing this library. I tried setting CURLOPT_VERBOSE first, but
that prints a bunch of debugging information in addition to the actual data.